+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fix orbits randomly showing up as white
 - Fix random position markers showing up
 - Fix rocket appearing out of orbit when zoomed on
+- Make rocket rotate around its axis
 
 # v0.3.0 - 2019-08-17
 - Use Moon orbital elements of 2019-08-11